+++ docs/source/changelog.md
@@ -2,6 +2,7 @@
 
 ## Latest
 - Allow invocation of undecorated task functions when calling task object outside of flow definition context.
+- Rename `ignore_fresh_input` to `ignore_cache_function`.
 - Fix race condition leading to `JSONDecodeError` in [](#ParquetTableCache) when setting `store_input: true` together with the `DaskEngine`.
 - Fix running subset of tasks not working due to tables and blobs being retrieved from wrong schema.